add_test_page_app&css

+++ accounts/models.py
@@ -1,10 +1,9 @@
 
+
+from django.contrib.auth.models import AbstractUser
 from django.db import models
 
 
-from django.contrib.auth.models import AbstractUser
-
-
 class CustomUser(AbstractUser):
-    age=models.PositiveBigIntegerField(null=True,blank=True)
+    age = models.PositiveIntegerField(null=True, blank=True)
